move around and rename some show related stuff
